refactor: 将 signal.js 移至 devices-dag 文件夹
