+++ src/core/devices-dag/index.js
@@ -13,3 +13,4 @@ export {
 export { DevicesDAGNode, DevicesDAGEdge } from "./dag-node-edge.js";
 export { DAGBuilder, DAGNodeBuilder, createSubDAG } from "./dag-builder.js";
 export { dagToString, dagToMermaid } from "./dag-debug.js";
+export { SignalPacket } from "./signal.js";
